fixing merge conflicts prepping for demo
