+++ src/app/admin/components/ReportsList.tsx
@@ -1,9 +1,7 @@
 'use client';
-import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import type { RouterOutputs } from '@/trpc/react';
 import { Flag } from 'lucide-react';
-import Link from 'next/dist/client/link';
 
 type Report = RouterOutputs['reports']['getAll'][number];
 
@@ -40,12 +38,12 @@ export function ReportsList({ reports }: ReportsListProps) {
                                 </p>
                                 <p className="text-xs text-gray-400">{new Date(report.createdAt!).toLocaleDateString()}</p>
                             </div>
-                            <Button
+                            {/*<Button
                                 size="sm"
                                 className="rounded-lg border-0 bg-orange-600/10 text-orange-400 hover:bg-orange-500/50 px-3 py-1"
                             >
                                 <Link href={`/admin/reports/${report.id}`}>Investigate</Link>
-                            </Button>
+                            </Button>*/}
                         </div>
                     ))}
                 </div>
