+++ src/app/dashboard/jamFinder/JamFinderClient.tsx
@@ -259,7 +259,7 @@ export default function JamFinderClient() {
                                         />
                                         {isModalOpen && selectedProject === project.id && (
                                             <ProjectModal
-                                                id = {project.id}
+                                                id={project.id}
                                                 title={project.title ?? 'Untitled Jam'}
                                                 subtitle={project.subtitle ?? ''}
                                                 starts={format(project.startDateTime ?? new Date(), 'MMM dd, yyyy')}
@@ -319,7 +319,7 @@ export default function JamFinderClient() {
                                         />
                                         {isModalOpen && selectedProject === project.id && (
                                             <ProjectModal
-                                                id = {project.id}
+                                                id={project.id}
                                                 title={project.title ?? 'Untitled Jam'}
                                                 subtitle={project.subtitle ?? ''}
                                                 starts={format(project.startDateTime ?? new Date(), 'MMM dd, yyyy')}
