+++ src/server/db/schemas/profiles.ts
@@ -3,7 +3,6 @@ import { relations, sql } from 'drizzle-orm';
 import { primaryKey } from 'drizzle-orm/sqlite-core';
 import { createTable } from '../schema-util';
 import { users } from './auth';
-import { projectRegistrations } from './project-registration';
 import { candidateProfilesToProjectInstances } from './projects';
 
 export const candidateProfiles = createTable('candidate_profile', (d) => ({
